can: make DLC_SHIFT dependent on HAL version
With stm32g4xx-hal-driver v1.2.3 the `DLC_SHIFT` changed from 16 to 0. But
there is no preprocessor symbol that can be accessed easily to figure out
the HAL version.

With the same HAL version the pre-processor symbol `USB_EP0StartXfer` removed
from the HAL, so use this as an indicator if we use a new HAL or and old
one.

This patch makes it easier to switch between different HAL versions. Once
we use the new HAL permanently this can be cleaned up.