Make_Rules.mk: filter out stm32g4xx_hal_timebase_tim_template from HAL
This patch makes development easier, it allows building against an external
stm32g4xx_hal_driver without further modification.

The stm32g4xx_hal_driver comes with the template
`stm32g4xx_hal_timebase_tim_template.c`. It cannot be compiled as is,
`HAL_TIM_MODULE_ENABLED` has to be enabled.

As we do not want or need this code, we filter it out of the list of code
to be compiled.